requirements.txt
Django==3.0.6
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0

django-environ==0.4.5
requests==2.23.0
mimesis==4.0.0